src/baselines/logic-rag/orignal_code/requirements.txt
torch>=1.9.0
sentence-transformers>=2.2.0
openai>=1.0.0
tqdm>=4.62.0
numpy>=1.21.0
backoff>=1.11.0
ratelimit>=2.2.1
python-dotenv>=0.19.0
colorama>=0.4.6 